+++ src/app/(fullscreen)/event/success/_components/step2.tsx
@@ -2,7 +2,7 @@
 
 import DetailContent from "@/components/detail-content";
 import { useSelectedCinemaStore } from "app/_stores/use-selected-cinema-store";
-import { useEventFormStore } from "app/_stores/useEventCreateForm";
+import { useEventFormStore } from "app/_stores/use-event-create-form";
 
 export default function Step2() {
   const { formData } = useEventFormStore();
