+++ src/app/(tabs)/notifications/page.tsx
@@ -11,7 +11,7 @@ export default function NotificationPage() {
   }, [list]);
 
   return (
-    <div className="min-h-screen text-white">
+    <div className="h-[calc(100vh-102px)] overflow-y-scroll text-white">
       <h1 className="title-1-semibold px-5 pt-6 pb-7.5">알림</h1>
       {list.length === 0 ? (
         <p className="text-center text-gray-500">알림이 없습니다.</p>
